validasi kategori
ketika di gunakan tidak bisa dihapus

+++ app/Http/Controllers/ExpenditureCategoryController.php
@@ -112,6 +112,10 @@ class ExpenditureCategoryController extends Controller
             return response()->json(['error' => 'Kategori pendapatan tidak ditemukan.'], 404);
         }
 
+        if(count($category->histories) > 0 || count($category->regulars) > 0){
+          return response()->json(['message' => 'Kategori sedang digunakan.'], 403);
+        }
+
         $category->delete();
 
         return response()->json(['success' => true]);
